requirements.txt
Django==3.2.8
django-debug-toolbar==3.2.2
beautifulsoup4==4.11.1
